Group pages with tabs and add custom user file
Improve overall grouping and architecture to mimic the way other LuCI modules are displayed.
Add custom user file to allow the user to map MAC addresses with usernames. This file can be edited in the UI.
Use current pathname, as a fail safe way, to normalize urls for the ajax calls in the static javascript.
Add `luci` and `wrtbwmon` as dependencies in the `control` file.

Closes #17, Closes #13